Check URI.equals() for URL equality
A server may return a Location header with a differently percentage encoded
value, e.g. whether to encode `[` and `]`. `rclone serve webdav` is one server
that has this behavior. Since we are already creating URI instances from
HttpUrl, we can just compare them to check for this kind of equality.